+++ libraries/resource_providers/api300/c7000/san_manager_provider.rb
@@ -1,4 +1,4 @@
-# (c) Copyright 2016-2017 Hewlett Packard Enterprise Development LP
+# (c) Copyright 2017 Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -9,18 +9,14 @@
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
 
-OneviewCookbook::ResourceBaseProperties.load(self)
+require_relative '../../api200/san_manager_provider'
 
-default_action :add
-
-action :add do
-  OneviewCookbook::Helper.do_resource_action(self, :SANManager, :add_or_edit)
-end
-
-action :add_if_missing do
-  OneviewCookbook::Helper.do_resource_action(self, :SANManager, :add_if_missing)
-end
-
-action :remove do
-  OneviewCookbook::Helper.do_resource_action(self, :SANManager, :remove)
+module OneviewCookbook
+  module API300
+    module C7000
+      # SANManager API300 C7000 provider
+      class SANManagerProvider < API200::SANManagerProvider
+      end
+    end
+  end
 end